requirements.txt
amqp==5.0.9
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.3
celery-progress==0.1.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.9
django-celery-results==2.2.0
django-environ==0.8.1
django-heroku==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.3
MarkupSafe==2.0.1
numpy==1.22.1
oauthlib==3.1.1
opencv-python-headless==4.5.5.62
packaging==21.3
Pillow==8.4.0
prompt-toolkit==3.0.28
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python3-openid==3.2.0
pytz==2021.3
redis==4.1.4
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.13.3
